fix migration

+++ src/database/migrations/2024_01_30_141000_create_industry_inputs_table.php
@@ -8,32 +8,36 @@ class CreateIndustryInputsTable extends Migration
 {
     public function up(): void
     {
-        Schema::create('pi_factory_inputs', function (Blueprint $table) {
-            $table->integer('schematic_id');
-            $table->integer('input_type_id');
-            $table->integer('quantity_consumed');
+        if (!Schema::hasTable('pi_factory_inputs')) {
+            Schema::create('pi_factory_inputs', function (Blueprint $table) {
+                $table->integer('schematic_id');
+                $table->integer('input_type_id');
+                $table->integer('quantity_consumed');
 
-            $table->foreign('schematic_id')
-                ->references('schematic_id')
-                ->on('universe_schematics')
-                ->cascadeOnDelete();
+                $table->foreign('schematic_id')
+                    ->references('schematic_id')
+                    ->on('universe_schematics')
+                    ->cascadeOnDelete();
 
-            $table->foreign('input_type_id')
-                ->references('typeID')
-                ->on('invTypes')
-                ->cascadeOnDelete();
+                $table->foreign('input_type_id')
+                    ->references('typeID')
+                    ->on('invTypes')
+                    ->cascadeOnDelete();
 
-            $table->primary(['schematic_id', 'input_type_id']);
-        });
+                $table->primary(['schematic_id', 'input_type_id']);
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('pi_factory_inputs', function (Blueprint $table) {
-            $table->dropForeign('pi_factory_inputs_schematic_id_foreign');
-            $table->dropForeign('pi_factory_inputs_input_type_id_foreign');
-        });
+        if (Schema::hasTable('pi_factory_inputs')) {
+            Schema::table('pi_factory_inputs', function (Blueprint $table) {
+                $table->dropForeign('pi_factory_inputs_schematic_id_foreign');
+                $table->dropForeign('pi_factory_inputs_input_type_id_foreign');
+            });
 
-        Schema::drop('pi_factory_inputs');
+            Schema::drop('pi_factory_inputs');
+        }
     }
 }
